% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey_to_matrix.R
\name{tukey_to_matrix}
\alias{tukey_to_matrix}
\title{Convert TukeyHSD() Output to a Matrix Array}
\usage{
tukey_to_matrix(tukey_object, element, significance = TRUE)
}
\arguments{
\item{tukey_object}{The input data. Should be a TukeyHSD() object. Named tukey_object since input should be a TukeyHSD() object.}

\item{element}{The element of the TukeyHSD() object to subset and convert. Does not need to be a string. Named element since this should be the element from the TukeyHSD() object that wants to be extracted.}

\item{significance}{Defaults to TRUE. Should statistically significant p-values be replaces with asterisks? Named significance since this controls if significant p-values are replaced with asterisks.}
}
\value{
A dataframe object with rows and columns of comparisons.
}
\description{
Create a user-readable matrix array from the output of TukeyHSD(). Takes the output of TukeyHSD() and converts the output object into a matrix array dataframe.
}
\examples{
AOV_Out <- aov(Petal.Length ~ Species, iris)
TukeyHSD_Out <- TukeyHSD(AOV_Out)
tukey_to_matrix(tukey_object = TukeyHSD_Out, element = Species)
tukey_to_matrix(TukeyHSD_Out, Species, significance = FALSE)
}
